:sparkles: [#7] 기본 경로를 app으로 이동
